finish product create

+++ src/Voucher/ConnectionPoint.Voucher.Application/Dtos/Discountable/CreateDiscountableDto.cs
@@ -6,7 +6,7 @@ namespace ConnectionPoint.Voucher.Application.Dtos.Discountable
     {
         public Guid DiscountableId { get; set; }
         public string DiscountableType { get; set; }
-        public double NetPrice { get; set; }
+        public decimal NetPrice { get; set; }
         public List<CouponDto> Coupons { get; set; } = new List<CouponDto>();
         public List<Guid> CouponIds { get; set; } = new List<Guid>();
